feat(instancing): Phase 2a step 5 — Renderer write/bind + ForwardPass variant selection
Renderer.h: forward-declared InstanceDataBuffer + unique_ptr member (avoids circular
include), m_instancedTransportEnabled flag + public IsInstancedTransportEnabled().

Renderer.cpp: Init() creates InstanceDataBuffer (1024-slot budget). Render(job):
after FeedUniforms, when flag on and scope guard passes (non-skinned, non-shader-
material), Write(0, perDrawUboLayout), Flush(), SetTexture(14, GetTexture()).

ForwardPass.cpp::RenderOpaque: vert->SetDefine(TK_INSTANCED, flag ? 1 : 0) before
both CreateProgram calls.

TextureBuffer.h: Resize now sets InternalFormat/Type/MinFilter/MagFilter correctly,
and calls m_buffer->Init(m_data.data()) — the runtime fix deferred from step 2.

instanceDataInc.shader: TK_INSTANCE_ID macro now casts to uint (gl_InstanceID is int
in GLSL ES 3.00; implicit int->uint does not exist). TK_INSTANCED=1 variant fixed.

Flag is still false — all draws take the legacy path (byte-identical).

+++ Resources/Engine/Shaders/instanceDataInc.shader
@@ -32,10 +32,12 @@
   // (which already includes the base instance); GL ES 3.00 exposes gl_InstanceID (which does
   // not). TK_INSTANCE_ID is the caller's spelling; the desktop-GL gl_InstanceID-excludes-base
   // gotcha is absorbed here.
+// gl_InstanceID (GLES) / gl_InstanceIndex (Vulkan) are int; LoadInstance expects uint.
+// Cast here so call sites don't each need one.
 #ifdef VULKAN
-  #define TK_INSTANCE_ID gl_InstanceIndex
+  #define TK_INSTANCE_ID uint(gl_InstanceIndex)
 #else
-  #define TK_INSTANCE_ID gl_InstanceID
+  #define TK_INSTANCE_ID uint(gl_InstanceID)
 #endif
 
   // Texel stride per instance, in RGBA32F texels. MUST match the C++ value
